added: telegram bot integration

+++ backend/inference/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 
 class Message(models.Model):
-    chat_id = models.IntegerField()
+    chat_id = models.BigIntegerField() 
     text = models.TextField()
     created_at = models.DateTimeField(auto_now_add=True)
     class_0_probability = models.FloatField(blank=True, null=True)
